APIs should not provide CLI help
