+++ cmd/check_process/cmd/root_linux.go
@@ -1,0 +1,8 @@
+// +build !windows
+
+package cmd
+
+func getHelpOsConstrained() string {
+	return `
+Note: Process names in POSIX systems are case sensitive.`
+}
