+++ cmd/check_service/cmd/root_linux.go
@@ -4,6 +4,15 @@ package cmd
 
 import "github.com/spf13/cobra"
 
+func getHelpOsConstrained() string {
+	return `
+
+For Linux, the only check done is for a running state. Both the --name (-n) and
+-manager (-m) options must be specified and the service is only checked
+to see if it is running.
+`
+}
+
 func addFlagsOsConstrained(cmd *cobra.Command) {
 	const managerFlag = "manager"
